+++ tests/dummy/app/controllers/application.js
@@ -1,8 +1,9 @@
 import Ember from 'ember';
 
 export default Ember.Controller.extend({
+  dynamicArgument: "argument 3",
   actions: {
-    save: function(callback) {
+    save: function(callback, passedArgument1, passedArgument2, passedArgument3) {
       var _this = this;
       var promise = new Ember.RSVP.Promise(function(resolve, reject) {
         if (_this.get('rejectPromise')) {
@@ -11,6 +12,9 @@ export default Ember.Controller.extend({
           }, _this.get('timeoutLength'));
         } else {
           Ember.run.later(function() {
+            _this.set('actionArgument1', passedArgument1);
+            _this.set('actionArgument2', passedArgument2);
+            _this.set('actionArgument3', passedArgument3);
             resolve();
           }, _this.get('timeoutLength'));
         }