+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 # master
 
+* Implements helper so that arguments passed in are passed to the action
+  called to fix
+[#27](https://github.com/dockyard/ember-cli-async-button/issues/27)
+[#29](https://github.com/dockyard/ember-cli-async-button/pull/29)
+* Updated addon for ember-cli 0.2.0-beta.1
+  [#28](https://github.com/dockyard/ember-cli-async-button/pull/28)
 * [DEPRECATION] use `fulfilled` instead of `resolved` [#9](https://github.com/dockyard/ember-cli-async-button/pull/9)
 * Updated addon for ember-cli 0.0.41 [#10](https://github.com/dockyard/ember-cli-async-button/pull/10)